+++ controllers/role.js
@@ -1,4 +1,5 @@
 const Model = require("../models/role")
+const User = require("../models/user")
 
 module.exports.getAll = (req, res, next) => {
   try {
@@ -66,19 +67,29 @@ module.exports.update = (req, res, next) => {
 
 
 }
-module.exports.delete = (req, res, next) => {
+module.exports.delete = async (req, res, next) => {
   let {id} = req.params
-  Model.findByIdAndDelete(id, (err, data) => {
-    if(err) console.log(err)
-    if(data){
-      res.status(200).json({
-        message: `Delete your data by id: ${id}`,
-        data
-      })
-    } else {
-      res.status(404).json({
-        message: `Data for ID: ${id} is not found, can't be deleted!`,
-      })
-    }
-  })
+  let checkDependency = await User.find({roles: id})
+  console.log(checkDependency)
+  if(!checkDependency.length){
+    Model.findByIdAndDelete(id, (err, data) => {
+      if(err) console.log(err)
+      if(data){
+        res.status(200).json({
+          message: `Удалена роль: ${data.name}`,
+          data
+        })
+      } else {
+        res.status(404).json({
+          message: `Не найдена роль ID: ${id}`,
+        })
+      }
+    })
+  } else {
+    res.status(409).json({
+      message: `Нельзя удалить. Имеются зависимости.`,
+      dependencies: checkDependency.map(item => {return item.name})
+    })
+  }
+
 }
